@@ -17,12 +17,11 @@
   },
 
   "require": {
-    "php": "^5.4 || ^7.0",
-
+    "php": "^7.2",
     "ext-json": "*"
   },
 
   "require-dev": {
-    "phpunit/phpunit": "~4.8"
+    "phpunit/phpunit": ">=8.5.8"
   }
 }
